falco-rules-viewer/backend/requirements.txt
Flask
PyYaml
Flask-Cors